+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## [unreleased]
 
+## 28.1.0
+
 ### Fixed
 
 Loose Constraint on first position didn't found all matched records.